refactor: rimuovere backward-compat code morto e allineare cross-repo
Fase 1 — Taglio netto (safe, 756 test OK):
- Eliminato mcp/contracts.py (shim modulo rinominato, zero consumatori)
- Eliminato block_url_direct alias in scaffold/sources.py
- Eliminato scaffold re-export in scout/__init__.py
- Rimosso cross_year response key duplicata in MCP schema_ops.py
- Rimosso build_profile_hints alias (usare sniff_source_file)
- Rimosso re-export run_context.py (import run_records direttamente)
- Eliminato _deprecated_cross_year_cmd CLI stub
- Eliminato inspect url deprecato (url_ops.py + registrazione)
- Rimosso MCP aliases toolkit_show_schema / toolkit_raw_profile (#299)

Fase 2 — Cross-repo fix:
- Eliminato lab-ops/scripts/scout_one.py (import cmd_scout_url rotto)
- profile_excel: reso pubblico, aggiornato source-observatory import

Totale: -114 nette su 22 file.
