+++ toolkit/cli/cmd_resume.py
@@ -9,7 +9,7 @@ from toolkit.core.config import load_config
 from toolkit.core.logging import get_logger
 from toolkit.core.metadata import read_layer_metadata
 from toolkit.core.paths import METADATA, layer_year_dir
-from toolkit.core.run_context import get_run_dir, latest_run, read_run_record
+from toolkit.core.run_records import get_run_dir, latest_run, read_run_record
 
 
 def _artifact_exists(path: Path) -> bool:
