+++ toolkit/cli/cmd_status.py
@@ -9,7 +9,7 @@ import typer
 from toolkit.cli.common import format_profile_preview, load_layer_profile_summaries
 from toolkit.core.config import load_config
 from toolkit.core.paths import METADATA, RAW_PROFILE_DIR, RAW_SUGGESTED_READ, layer_dataset_dir
-from toolkit.core.run_context import get_run_dir, read_run_record
+from toolkit.core.run_records import get_run_dir, read_run_record
 from toolkit.cli.inspect.readiness_ops import summary as _summary
