+++ toolkit/cli/inspect/__init__.py
@@ -8,7 +8,6 @@ from toolkit.cli.inspect.profile_ops import profile
 from toolkit.cli.inspect.schema_diff_ops import schema_diff
 from toolkit.cli.inspect.schema_ops import schema
 from toolkit.cli.inspect.probe_ops import probe
-from toolkit.cli.inspect.url_ops import url
 
 
 def register(app: typer.Typer) -> None:
@@ -19,5 +18,4 @@ def register(app: typer.Typer) -> None:
     inspect_app.command("schema-diff")(schema_diff)
     inspect_app.command("schema")(schema)
     inspect_app.command("probe")(probe)
-    inspect_app.command("url", hidden=True, help="[DEPRECATO] Usa 'toolkit scout'.")(url)
     app.add_typer(inspect_app, name="inspect", help="Ispeziona path, schema, readiness e URL del dataset.")
