+++ toolkit/cli/inspect/_helpers.py
@@ -23,7 +23,7 @@ from toolkit.core.paths import (
     MART_VALIDATION,
     layer_year_dir,
 )
-from toolkit.core.run_context import get_run_dir, latest_run
+from toolkit.core.run_records import get_run_dir, latest_run
 from toolkit.core.support import resolve_support_payloads
 from toolkit.profile.raw import sniff_source_file
